services/order/requirements.txt
fastapi
uvicorn[standard]
pydantic
prometheus-fastapi-instrumentator
aiokafka
asyncpg
sqlalchemy[asyncio]
python-jose[cryptography]

